Consolidate Benchmarks projects
