--- docs/trace/benchmark/benchmark.csproj
@@ -1,9 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-  <ItemGroup>
-    <!---
-    <PackageReference Include="OpenTelemetry.Exporter.Console" Version="$(OpenTelemetryPkgVer)" />
-    -->
-    <PackageReference Include="BenchmarkDotNet" Version="$(BenchmarkDotNetPkgVer)" />
-    <ProjectReference Include="$(RepoRoot)\src\OpenTelemetry\OpenTelemetry.csproj" />
-  </ItemGroup>
-</Project>
