+++ test/Benchmarks/README.md
@@ -3,7 +3,7 @@
 Use the following example to run Benchmarks from command line:
 (change parameters as necessary)
 
-Navigate to `./test/benchmarks` directory and run the following command:
+Navigate to `./test/Benchmarks` directory and run the following command:
 
 `dotnet run --framework netcoreapp3.1 --configuration Release --filter
-*OpenTelemetrySdkBenchmarksActivity*`
+*TraceBenchmarks*`
