+++ test/Benchmarks/Tracing/SpanCreationScenarios.cs
@@ -15,9 +15,8 @@
 // </copyright>
 
 using System.Diagnostics;
-using OpenTelemetry.Trace;
 
-namespace Benchmarks.Tracing
+namespace OpenTelemetry.Trace.Benchmarks
 {
     internal class SpanCreationScenarios
     {
